apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
slowapi==0.1.9
asyncpg==0.30.0
httpx==0.28.0
python-dotenv==1.0.1
jsonschema==4.23.0
pydantic==2.10.2
pydantic-settings==2.6.1
beautifulsoup4==4.12.3
